+++ state-root-storage-foreign/test/utils/TestUtils.sol
@@ -112,7 +112,7 @@ abstract contract TestUtils is Test, BLSUtilsFFI, BLSMockAVSDeployer {
         return (signatoryRecordHash, nonSignerStakesAndSignature);
     }
 
-    function quorumThreshold(uint256 denominator, uint256 nonSignerCount) public view returns (uint32) {
+    function quorumThreshold(uint32 denominator, uint256 nonSignerCount) public view returns (uint32) {
         return uint32((maxOperatorsToRegister - nonSignerCount) * denominator / maxOperatorsToRegister);
     }
 }
